Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-float num1;
+double num1;
 scanf("%lf",&num1);
 printf("You entered:%.2lf\n",num1);
 return 0;
